Was using int instead of size_t for GetChildCount, vector returns a size_t when using .size() so just changed the function to return size_t too

+++ Minigin/GameObject.cpp
@@ -77,7 +77,7 @@ void dae::GameObject::SetParent(GameObject* parent, bool keepWorldPosition)
 	}
 }
 
-int dae::GameObject::GetChildCount() const
+size_t dae::GameObject::GetChildCount() const
 {
 	return m_children.size();
 }
